Movement
Added comments.

Added functionality for transition between aerial and ground view (no teleporting while staying in the air yet, need to add a button for it).

I think you can change the teleport to a straight line by changing the projectile velocity (I think that is what the variable is called).